Delete LN addresses on store delete
fixes #3596
