+++ BTCPayServer/Controllers/UILNURLController.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Globalization;
@@ -26,7 +25,6 @@ using BTCPayServer.Services.Stores;
 using LNURL;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.ModelBinding;
 using Microsoft.AspNetCore.Routing;
 using NBitcoin;
 using NBitcoin.Crypto;
@@ -147,35 +145,6 @@ namespace BTCPayServer
 
             public EditLightningAddressItem Add { get; set; }
             public List<EditLightningAddressItem> Items { get; set; } = new List<EditLightningAddressItem>();
-        }
-
-        public class LightningAddressSettings
-        {
-            public class LightningAddressItem
-            {
-                public string StoreId { get; set; }
-                [Display(Name = "Invoice currency")]
-                public string CurrencyCode { get; set; }
-
-                [Display(Name = "Min sats")]
-                [Range(1, double.PositiveInfinity)]
-                public decimal? Min { get; set; }
-
-                [Display(Name = "Max sats")]
-                [Range(1, double.PositiveInfinity)]
-                public decimal? Max { get; set; }
-            }
-
-            public ConcurrentDictionary<string, LightningAddressItem> Items { get; set; } =
-                new ConcurrentDictionary<string, LightningAddressItem>();
-
-            public ConcurrentDictionary<string, string[]> StoreToItemMap { get; set; } =
-                new ConcurrentDictionary<string, string[]>();
-
-            public override string ToString()
-            {
-                return null;
-            }
         }
 
         private async Task<LightningAddressSettings> GetSettings()
